# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/channels/{channelId}/messages": {
      "post": {
        "operationId": "MessagesController_send",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Messages"
        ]
      },
      "get": {
        "operationId": "MessagesController_findAll",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Messages"
        ]
      }
    },
    "/channels/{channelId}/messages/{messageId}": {
      "delete": {
        "operationId": "MessagesController_remove",
        "parameters": [
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Messages"
        ]
      }
    },
    "/channels/{channelId}/messages/{messageId}/pin": {
      "patch": {
        "operationId": "MessagesController_pin",
        "parameters": [
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Messages"
        ]
      }
    },
...
    "/private-chats/{privateChatId}/messages": {
      "post": {
        "operationId": "PrivateChatMessagesController_send",
        "parameters": [
          {
            "name": "privateChatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "PrivateChatMessages"
        ]
      },
      "get": {
        "operationId": "PrivateChatMessagesController_list",
        "parameters": [
          {
            "name": "privateChatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PrivateChatMessages"
        ]
      }
    },
...
    "/groups/{groupId}/private-chats": {
      "post": {
        "operationId": "PrivateChatsController_create",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "PrivateChats"
        ]
      },
      "get": {
        "operationId": "PrivateChatsController_findAll",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PrivateChats"
        ]
      }
    },
    "/groups/{groupId}/private-chats/{privateChatId}": {
      "delete": {
        "operationId": "PrivateChatsController_remove",
        "parameters": [
          {
            "name": "privateChatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PrivateChats"
        ]
      }
    }
  },
  "info": {
    "title": "LECTERN",
    "description": "Backend API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateMessageDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}